+++ src/test.rs
@@ -1,0 +1,4 @@
+//! Testing.
+
+mod basic;
+mod collect;