+++ CHANGES.md
@@ -1,3 +1,9 @@
+v1.2.0
+
+- cleaned `lazy_static` re-export, now only re-exporting `lazy_static` macro
+- added a `capacity` method on the consign
+- added `collect`, `shrink_to_fit`, `collect_to_fit` and `reserve` to the `HashConsign` trait
+
 v1.1.0
 
 - removed systematic unsafe implementations of `Send` and `Sync` for `HConsed`,